+++ academy-ui/src/app/api/s3-signed-urls/route.ts
@@ -1,11 +1,11 @@
-import { MutationCreateSignedUrlArgs } from '@/graphql/generated/generated-types';
-import { getSpaceById } from '@/app/api/helpers/space/getSpaceById';
 import { logError } from '@/app/api/helpers/adapters/errorLogger';
+import { withErrorHandling } from '@/app/api/helpers/middlewares/withErrorHandling';
 import { presignedUrlCreator } from '@/app/api/helpers/s3/getPresignedUrl';
 import { checkEditSpacePermission } from '@/app/api/helpers/space/checkEditSpacePermission';
-import { withErrorHandling } from '@/app/api/helpers/middlewares/withErrorHandling';
+import { getSpaceById } from '@/app/api/helpers/space/getSpaceById';
+import { validateApiKey } from '@/app/api/helpers/validateApiKey';
+import { MutationCreateSignedUrlArgs } from '@/graphql/generated/generated-types';
 import { NextRequest, NextResponse } from 'next/server';
-import { prisma } from '@/prisma';
 
 async function postHandler(req: NextRequest) {
   const args: MutationCreateSignedUrlArgs = await req.json();
@@ -15,39 +15,15 @@ async function postHandler(req: NextRequest) {
   try {
     if (apiKey) {
       await validateApiKey(apiKey, args.spaceId);
-    } else if (spaceById.id !== 'tidbitshub') {
+    } else {
       await checkEditSpacePermission(spaceById, req);
     }
-
     const url = await presignedUrlCreator.createSignedUrl(spaceById.id, args.input);
     return NextResponse.json({ url }, { status: 200 });
   } catch (error) {
-    await logError((error as any)?.response?.data || 'Error in createSignedUrls', {}, error, null, null);
+    await logError((error as any)?.response?.data || 'Error in createSignedUrls', {}, error as any, null, null);
     throw error;
   }
-}
-
-async function validateApiKey(apiKey: string, spaceId: string) {
-  const spaceIntegration = await prisma.spaceIntegration.findUnique({
-    where: { spaceId },
-  });
-
-  if (!spaceIntegration?.spaceApiKeys) {
-    throw new Error('API Key is not valid');
-  }
-
-  const apiKeyEntry = spaceIntegration.spaceApiKeys.find((key) => key.apiKey === apiKey);
-
-  if (!apiKeyEntry) {
-    throw new Error('API Key is not valid');
-  }
-
-  apiKeyEntry.lastUsed = new Date();
-
-  await prisma.spaceIntegration.update({
-    where: { spaceId },
-    data: { spaceApiKeys: spaceIntegration.spaceApiKeys },
-  });
 }
 
 export const POST = withErrorHandling(postHandler);
